requirements.txt
flask==2.0.1
googletrans==3.1.0a0
gTTS==2.2.3
python-docx==0.8.11